+++ tests/testthat/test-read-lines.R
@@ -59,6 +59,17 @@ test_that("allocation works as expected", {
   expect_equal(length(read_lines(tmp)), 2^11)
 })
 
+test_that("read_lines(skip_empty_rows) works when blank lines are at the end of the file (#968)", {
+  tmp <- tempfile()
+  on.exit(unlink(tmp))
+
+  writeLines(con = tmp,
+"test
+")
+
+  expect_equal(read_lines(tmp, skip_empty_rows = TRUE), "test")
+})
+
 # These tests are slow so are commented out
 #test_that("long vectors are supported", {
   #tmp <- tempfile(fileext = ".gz")